my::Font::GetHeight, SetHeight
git-svn-id: svn://127.0.0.1/repos/trunk@383 a5c22d9f-9f17-4845-9224-d5bed68bb274

+++ myd3dlib/src/myUi.cpp
@@ -507,7 +507,7 @@ void EditBox::Draw(UIRender * ui_render, float fElapsedTime, const Vector2 & Off
 
 				if(!m_bInsertMode)
 				{
-					int charWidth;
+					float charWidth;
 					if(m_nCaret < (int)m_Text.length())
 					{
 						const Font::CharacterInfo & info = Skin->m_Font->GetCharacterInfo(m_Text[m_nCaret]);
